Provide legal license contract template
This patch provides a template for the COPYRIGHT file (a.k.a. LICENSE
file or COPYING file), which contains the license agreement to use,
copy, modify, distribute, &c. the Software (defined as the code and the
documentation).

It is a very permissive license, lacking warranty of any kind, similar
to BSD licenses.  The text is enough to comply with International
Copyright laws (accepted in almost all legislation worldwides).

Signed-off-by: Nicolas Paul <n@nc0.fr>